+++ js/constants.js
@@ -1,0 +1,35 @@
+const ADVERT_COUNT = 10;
+const ADVERT_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+const CHECKIN_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const CHECKOUT_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES_ADVERT = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const PHOTOS_ADVERT = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+const RANDOM_ADVERTS = [];
+
+export {ADVERT_COUNT, ADVERT_TYPES, CHECKIN_TIME, CHECKOUT_TIME, FEATURES_ADVERT,
+  PHOTOS_ADVERT, RANDOM_ADVERTS};
